Improve instrument-specific key signatures example
I think this small change to the trumpet/tuba with 2 different key signatures example might better illustrate how it could be interesting to write a piece of music where the instruments are playing in different keys. The difference is subtle, but if you play `keysig1.alda` and `keysig2.alda` back to back, you can hear the difference in character.

I also bumped the tuba down a couple of octaves, which makes it a little easier for the listener to tell the two instruments apart.